Feature

+++ src/main/java/com/example/AnimeAPI/service/AnimeService.java
@@ -3,10 +3,17 @@ package com.example.AnimeAPI.service;
 import com.example.AnimeAPI.exception.InformationExistException;
 import com.example.AnimeAPI.exception.InformationNotFoundException;
 import com.example.AnimeAPI.model.Anime;
+
+import com.example.AnimeAPI.model.User;
+import com.example.AnimeAPI.repository.AnimeRepository;
+import com.example.AnimeAPI.security.MyUserDetails;
+
 import com.example.AnimeAPI.model.Genre;
 import com.example.AnimeAPI.repository.AnimeRepository;
 import com.example.AnimeAPI.repository.GenreRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,6 +29,11 @@ public class AnimeService {
     public AnimeService(AnimeRepository animeRepository, GenreRepository genreRepository){
         this.animeRepository = animeRepository;
         this.genreRepository = genreRepository;
+    }
+
+    public static User getCurrentLoggedInUser() {
+        MyUserDetails userDetails = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return userDetails.getUser();
     }
 
     /**
@@ -42,13 +54,16 @@ public class AnimeService {
      * @throws InformationNotFoundException if the anime already exists.
      */
     public Anime createAnime(Anime animeObject){
-        Optional<Anime> anime = animeRepository.findByTitle(animeObject.getTitle());
-        if(anime.isPresent()){
-            throw new InformationNotFoundException("This anime already exists:" + animeObject.getTitle());
-        }else {
-            return animeRepository.save(animeObject);
+        User user = getCurrentLoggedInUser();
+        if (user.getUserType().equals("ADMIN")) {
+            Optional<Anime> anime = animeRepository.findByTitle(animeObject.getTitle());
+            if(anime.isPresent()){
+                throw new InformationNotFoundException("This anime already exists:" + animeObject.getTitle());
+            }else {
+                return animeRepository.save(animeObject);
+            }
         }
-
+        return null;
     }
 
     /**
@@ -61,13 +76,17 @@ public class AnimeService {
      * @return Anime {Object}
      */
     public Anime deleteAnime(Long animeId) {
-        Optional<Anime> anime = animeRepository.findById(animeId);
-        if (anime.isPresent()) {
-            animeRepository.delete(anime.get());
-            return anime.get();
-        } else {
-            throw new InformationNotFoundException("Anime with given id " + animeId + " does not exist.");
+        User user = getCurrentLoggedInUser();
+        if (user.getUserType().equals("ADMIN")) {
+            Optional<Anime> anime = animeRepository.findById(animeId);
+            if (anime.isPresent()) {
+                animeRepository.delete(anime.get());
+                return anime.get();
+            } else {
+                throw new InformationNotFoundException("Anime with given id " + animeId + " does not exist.");
+            }
         }
+        return null;
     }
 
     /**
